# back-end/documents/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the server for the game Beat My Wordle.",
    "version": "1.0.0",
    "title": "Beat My Wordle"
  },
  "host": "wordle.keyuka.ca",
  "basePath": "/1",
  "tags": [
    {
      "name": "users",
      "description": "User management and authentication"
    },
    {
      "name": "scores",
      "description": "For game score related queries"
    },
    {
      "name": "words",
      "description": "For user word entry related queries"
    },
    {
      "name": "games",
      "description": "For being matched with another player's word"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "General login to the application",
        "description": "",
        "operationId": "generalLogin",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Users object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 OK"
          },
          "403": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/users/signup": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "General user signup",
        "operationId": "userSignup",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Users object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 OK"
          },
          "403": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/users/admin/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Admin login",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Users object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 OK"
          }
        }
      }
    },
    "/words/check": {
      "get": {
        "tags": [
          "words"
        ],
        "summary": "Checks user submitted 5 letter word against an external dictionary API",
        "description": "Checks a single word to see if it a valid 5 letter word",
        "operationId": "wordCheck",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "word",
            "in": "query",
            "description": "Word that needs to be checked by dictionary",
            "required": true,
            "type":"string"
          }
        ],
        "responses": {
          "200": {
            "description": "Word was successfully checked"
          },
          "404": {
            "description": "Word could not be checked, endpoint not reached"
          }
        }
      }
    },
    "/words/upload": {
      "put": {
        "tags": [
          "words"
        ],
        "summary": "Upload or update a checked word to the Word microservice",
        "description": "Send valid word to word microservice ",
        "operationId": "wordUpload",
        "produces": ["html/text"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Username and their uploaded word",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Words"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Word was successfully added"
          },
          "403": {
            "description": "Unauthorized."
          },
          "400": {
            "description": "Bad request."
          },
          "500": {
            "description": "Word could not be added, could not reach endpoint"
          }
        }
      }
    },
    "/words/{username}": {
      "delete": {
        "tags": [
          "words"
        ],
        "summary": "Delete the word hosted by the user",
        "description": "Delete a word if a hosted ",
        "operationId": "wordDelete",
        "produces": ["html/text"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username associated to the word to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Word deleted successfully"
          },
          "404": {
            "description": "Username not associated to a word"
          }
        }
      }
    },
    "/profile/{username}": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "See the profile information of a user",
        "description": "Review the profile information of a user",
        "operationId": "viewProfile",
        "produces": ["html/text"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username associated to the user being viewed",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "User does not exist"
          }
        }
      }
    },
    "/scores/{username}": {
      "get": {
        "tags": [
          "scores"
        ],
        "summary": "Sees the score of a user",
        "description": "See the score a single user",
        "operationId": "scoreViewSingle",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username of the score to view ",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/scores/" : {
      "get": {
        "tags": [
          "scores"
        ],
        "summary": "Returns a map of all the scores stored",
        "description": "See all the user scores",
        "operationId": "scoreViewAll",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Scores not found"
          }
        }
      }
    },

    "/games": {
      "get": {
        "tags": [
          "matchmaking",
          "games"
        ],
        "summary": "Get a game",
        "parameters": [
          {
            "name": "word",
            "in": "query",
            "description": "Create a matchmade game",
            "required": true,
            "type":"string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 OK"
          },
          "404": {
            "description": "404 NO GAMES AVAILABLE"
          }
        }
      }
    }
  },
  "definitions": {
    "Users": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Users"
      }
    },
    "Words" : {
      "type": "object",
      "properties": {
        "username": {
          "type":"string"
        },
        "word": {
          "type":"string"
        }
      },
      "xml": {
        "name": "Words"
      }
    }
  }
}